meme_games/common/__init__.py
from fasthtml.common import *
from .lobby import *
from .notify import *
from .user import *
from .database import *
from .utils import *

rt = APIRouter()

#TODO fix imports, since common depends on init, which depends on common